+++ src/main/java/com/teamtea/eclipticseasons/api/data/season/definition/selector/IChangeSelector.java
@@ -19,24 +19,35 @@ import java.util.Optional;
 
 public interface IChangeSelector {
 
+    //@SuppressWarnings("unchecked")
+    //Codec<IChangeSelector> CCODEC =
+    //        Codec.either(
+    //                        BlockSelector.CODEC.codec(),
+    //                        Codec.STRING
+    //                                .xmap(s -> s.contains(":") ? ResourceLocation.parse(s) : EclipticSeasons.rl(s),
+    //                                        r -> r.getNamespace().equals(EclipticSeasonsApi.MODID) ? r.getPath() : r.toString())
+    //                                .dispatch("type", IChangeSelector::getType, id -> (MapCodec<IChangeSelector>) ChangeSelectors.CONDITIONS.get(id)))
+    //                .xmap(e -> e.left().map((bs -> (IChangeSelector) bs)).orElseGet(() -> e.right().orElseThrow()),
+    //                        ics -> ics instanceof BlockSelector bs ? Either.left(bs) : Either.right(ics));
     @SuppressWarnings("unchecked")
     Codec<IChangeSelector> CCODEC =
             Codec.either(
-                            BlockSelector.CODEC.codec(),
+
                             Codec.STRING
                                     .xmap(s -> s.contains(":") ? ResourceLocation.parse(s) : EclipticSeasons.rl(s),
                                             r -> r.getNamespace().equals(EclipticSeasonsApi.MODID) ? r.getPath() : r.toString())
-                                    .dispatch("type", IChangeSelector::getType, id -> (MapCodec<IChangeSelector>) ChangeSelectors.CONDITIONS.get(id)))
-                    .xmap(e -> e.left().map((bs -> (IChangeSelector) bs)).orElseGet(() -> e.right().orElseThrow()),
-                            ics -> ics instanceof BlockSelector bs ? Either.left(bs) : Either.right(ics));
+                                    .dispatch("type", IChangeSelector::getType, id -> (MapCodec<IChangeSelector>) ChangeSelectors.CONDITIONS.get(id)),
+                            BlockSelector.CODEC.codec())
+                    .xmap(e -> e.right().map((bs -> (IChangeSelector) bs)).orElseGet(() -> e.left().orElseThrow()),
+                            ics -> ics instanceof BlockSelector bs ? Either.right(bs) : Either.left(ics));
 
-    int DEFAULT_WEIGHT=10;
+    int DEFAULT_WEIGHT = 10;
 
     ResourceLocation getType();
 
     MapCodec<? extends IChangeSelector> codec();
 
-    boolean place(ServerLevel level, BlockPos pos, ISeasonChangeContext context);
+    boolean place(ServerLevel level, BlockPos origin, ISeasonChangeContext context);
 
 
     default int getWeight() {