+++ src/main/java/com/teamtea/eclipticseasons/api/data/season/definition/selector/MultiBlockSelector.java
@@ -59,13 +59,13 @@ public class MultiBlockSelector implements IChangeSelector {
     }
 
     @Override
-    public boolean place(ServerLevel level, BlockPos pos, ISeasonChangeContext context) {
-        List<MultiBlockSelector.Part> get = multiBlocks;
+    public boolean place(ServerLevel level, BlockPos origin, ISeasonChangeContext context) {
+        List<Part> get = multiBlocks;
         int count = 0;
         for (int j = 0, getSize = get.size(); j < getSize; j++) {
-            MultiBlockSelector.Part part = get.get(j);
+            Part part = get.get(j);
             BlockState newState = part.state;
-            BlockPos newpos = part.offset.map(pos::offset).orElse(pos);
+            BlockPos newpos = part.offset.map(origin::offset).orElse(origin);
             if (part.replace.isEmpty()) {
                 BlockState oldState = level.getBlockState(newpos);
                 if (oldState.canBeReplaced(Fluids.EMPTY)) {