+++ src/main/java/com/teamtea/eclipticseasons/common/core/crop/NaturalPlantHandler.java
@@ -145,7 +145,7 @@ public class NaturalPlantHandler {
                                     var blockStatePlaced = selectorsed.get(j);
                                     if (!blockStatePlaced.shouldApply(level, pos, context)) continue;
                                     weightIndex -= blockStatePlaced.getWeight();
-                                    if (weightIndex <= 0) {
+                                    if (weightIndex < 0) {
                                         chosen = blockStatePlaced;
                                         break;
                                     }