+++ src/main/java/com/teamtea/eclipticseasons/common/network/ESConfigFilePayload.java
@@ -1,0 +1,25 @@
+package com.teamtea.eclipticseasons.common.network;
+
+import com.teamtea.eclipticseasons.api.EclipticSeasonsApi;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.codec.ByteBufCodecs;
+import net.minecraft.network.codec.StreamCodec;
+import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
+import net.minecraft.resources.ResourceLocation;
+import net.neoforged.neoforge.network.codec.NeoForgeStreamCodecs;
+import org.jetbrains.annotations.NotNull;
+
+public record ESConfigFilePayload(String fileName, byte[] contents) implements CustomPacketPayload {
+    public static final Type<ESConfigFilePayload> TYPE = new Type<>(ResourceLocation.fromNamespaceAndPath(EclipticSeasonsApi.MODID, "config_file"));
+    public static final StreamCodec<FriendlyByteBuf, ESConfigFilePayload> STREAM_CODEC = StreamCodec.composite(
+            ByteBufCodecs.STRING_UTF8,
+            ESConfigFilePayload::fileName,
+            NeoForgeStreamCodecs.UNBOUNDED_BYTE_ARRAY,
+            ESConfigFilePayload::contents,
+            ESConfigFilePayload::new);
+
+    @Override
+    public @NotNull Type<ESConfigFilePayload> type() {
+        return TYPE;
+    }
+}