+++ src/main/java/com/teamtea/eclipticseasons/common/registry/SeasonDefinitionRegistry.java
@@ -7,6 +7,7 @@ import com.teamtea.eclipticseasons.api.data.season.definition.ChangeMode;
 import com.teamtea.eclipticseasons.api.data.season.definition.SeasonDefinition;
 import com.teamtea.eclipticseasons.api.data.season.definition.condition.EmptyAboveCondition;
 import com.teamtea.eclipticseasons.api.data.season.definition.selector.BlockSelector;
+import com.teamtea.eclipticseasons.api.data.season.definition.selector.FeatureSelector;
 import com.teamtea.eclipticseasons.api.data.season.definition.selector.MultiBlockSelector;
 import net.minecraft.advancements.critereon.BlockPredicate;
 import net.minecraft.advancements.critereon.StatePropertiesPredicate;
@@ -14,11 +15,13 @@ import net.minecraft.core.HolderSet;
 import net.minecraft.core.Vec3i;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
+import net.minecraft.data.worldgen.features.VegetationFeatures;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.biome.Biomes;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.DoublePlantBlock;
+import net.minecraft.world.level.block.StemBlock;
 import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
 
 import java.util.List;
@@ -33,8 +36,9 @@ public class SeasonDefinitionRegistry {
 
     public static void bootstrap2(BootstrapContext<SeasonDefinition> context) {
         var holderGetter = context.lookup(Registries.BIOME);
+        var placedFeatureHolderGetter = context.lookup(Registries.CONFIGURED_FEATURE);
 
-        HolderSet.Direct<Biome> plains = HolderSet.direct(holderGetter.getOrThrow(Biomes.THE_VOID));
+        HolderSet.Direct<Biome> plains = HolderSet.direct(holderGetter.getOrThrow(Biomes.PLAINS));
         Vec3i above = new Vec3i(0, 1, 0);
         List<EmptyAboveCondition> condition = List.of(EmptyAboveCondition.builder().above(true).build());
         context.register(test, new SeasonDefinition(
@@ -45,7 +49,20 @@ public class SeasonDefinitionRegistry {
                                         .original(BlockPredicate.Builder.block().of(Blocks.GRASS_BLOCK).build())
                                         .fixedSeed(true)
                                         .chance(1 / 16f)
-                                        .selector(BlockSelector.builder().conditions(condition).state(Optional.of(Blocks.SHORT_GRASS.defaultBlockState())).weight(22).offset(Optional.of(above)).build())
+                                        .selector(BlockSelector.builder().conditions(condition)
+                                                .state(Optional.of(Blocks.SHORT_GRASS.defaultBlockState())).weight(22).offset(Optional.of(above)).build())
+                                        .selector(MultiBlockSelector.builder()
+                                                .conditions(condition)
+                                                .multiBlock(MultiBlockSelector.Part.builder()
+                                                        .state(Blocks.PUMPKIN_STEM.defaultBlockState().setValue(StemBlock.AGE, StemBlock.MAX_AGE))
+                                                        .build())
+                                                .multiBlock(MultiBlockSelector.Part.builder()
+                                                        .state(Blocks.PUMPKIN.defaultBlockState())
+                                                        .build())
+                                                .offset(Optional.of(above))
+                                                .build())
+                                        .selector(FeatureSelector.builder().conditions(condition).feature(placedFeatureHolderGetter.getOrThrow(VegetationFeatures.PATCH_PUMPKIN)).weight(1).offset(Optional.of(above)).build())
+                                        .selector(FeatureSelector.builder().conditions(condition).feature(placedFeatureHolderGetter.getOrThrow(VegetationFeatures.BAMBOO_VEGETATION)).weight(1).offset(Optional.of(above)).build())
                                         .selector(BlockSelector.builder().conditions(condition).state(Optional.of(Blocks.DANDELION.defaultBlockState())).weight(1).offset(Optional.of(above)).build())
                                         .selector(BlockSelector.builder().conditions(condition).state(Optional.of(Blocks.OXEYE_DAISY.defaultBlockState())).weight(1).offset(Optional.of(above)).build())
                                         .build()