+++ src/main/java/com/teamtea/eclipticseasons/config/CommonConfig.java
@@ -19,7 +19,7 @@ import java.util.*;
 import java.util.stream.Stream;
 
 public class CommonConfig {
-    public static final ModConfigSpec COMMON_CONFIG = new ModConfigSpec.Builder().configure(com.teamtea.eclipticseasons.config.CommonConfig::new).getRight();
+    public static final ModConfigSpec COMMON_CONFIG = new ModConfigSpec.Builder().configure(CommonConfig::new).getRight();
 
     protected CommonConfig(ModConfigSpec.Builder builder) {
         Season.load(builder);
@@ -43,9 +43,18 @@ public class CommonConfig {
         public static ModConfigSpec.BooleanValue disableUniqueRebindingBiomeTags;
 
         public static ModConfigSpec.BooleanValue disableIceOrSnowCauldron;
+        public static ModConfigSpec.BooleanValue forceServerConfig;
 
         private static void load(ModConfigSpec.Builder builder) {
             builder.push("Debug");
+
+            forceServerConfig = builder
+                    .comment(
+                            "Force using server synchronized config when joining a multiplayer server.",
+                            "If disabled, client keeps its own local config when exit."
+                    )
+                    .define("ForceServerConfig", true);
+
             logIllegalUse = builder.comment("Log errors when internal functions are used incorrectly.")
                     .define("LogIllegalUse", false);
             notLightAbove = builder.comment("Prevent snow overlays from rendering under blocks with zero light emission.")
@@ -62,6 +71,7 @@ public class CommonConfig {
 
             disableIceOrSnowCauldron = builder.comment("Cauldrons will no longer collect snow or ice during winter weather.")
                     .define("DisableIceOrSnowCauldron", false);
+
 
             builder.pop();
         }
@@ -324,12 +334,12 @@ public class CommonConfig {
     public static boolean validSolarTerm(Object o) {
         if (o instanceof String s) {
             try {
-                com.teamtea.eclipticseasons.api.constant.solar.SolarTerm.valueOf(s);
+                SolarTerm.valueOf(s);
                 return true;
             } catch (IllegalArgumentException ignored) {
             }
         }
-        return o instanceof com.teamtea.eclipticseasons.api.constant.solar.SolarTerm;
+        return o instanceof SolarTerm;
     }
 
     public static Set<com.teamtea.eclipticseasons.api.constant.solar.Season> castSeasons(List<? extends String> strings) {
@@ -348,18 +358,18 @@ public class CommonConfig {
         return es1;
     }
 
-    public static Set<com.teamtea.eclipticseasons.api.constant.solar.SolarTerm> castSolarTerms(List<? extends String> strings) {
-        var es1 = EnumSet.noneOf(com.teamtea.eclipticseasons.api.constant.solar.SolarTerm.class);
+    public static Set<SolarTerm> castSolarTerms(List<? extends String> strings) {
+        var es1 = EnumSet.noneOf(SolarTerm.class);
         for (String string : strings) {
-            es1.add(com.teamtea.eclipticseasons.api.constant.solar.SolarTerm.valueOf(string));
+            es1.add(SolarTerm.valueOf(string));
         }
         return es1;
     }
 
-    public static List<com.teamtea.eclipticseasons.api.constant.solar.SolarTerm> castSolarTermList(List<? extends String> strings) {
-        var es1 = new ArrayList<com.teamtea.eclipticseasons.api.constant.solar.SolarTerm>();
+    public static List<SolarTerm> castSolarTermList(List<? extends String> strings) {
+        var es1 = new ArrayList<SolarTerm>();
         for (String string : strings) {
-            es1.add(com.teamtea.eclipticseasons.api.constant.solar.SolarTerm.valueOf(string));
+            es1.add(SolarTerm.valueOf(string));
         }
         return es1;
     }
@@ -497,7 +507,7 @@ public class CommonConfig {
             builder.push("Resource");
 
             RainTogether = builder.comment("Synchronizes weather states across all Overworld biomes, ensuring global rainfall.")
-                    .define("RainTogether", false);
+                    .define("RainTogether", true);
 
             SnowTogether = builder.comment("Synchronizes the snowfall schedule for all Overworld biomes.")
                     .define("SnowTogether", false);